COMP: fix couple of warnings (conversion size_t to int). COMP: OK with shared libraries on windows for vtk and itk

+++ Code/ExternalCode/MegaVTK/vtkRenderingAddOn/vtkViewImage2DCollection.cxx
@@ -352,7 +352,7 @@ void
 vtkViewImage2DCollection::
 SetSplinePlaneActorsVisibility( bool iVisibility )
 {
-	int numberOfActors = this->SlicePlaneActors.size();
+	size_t numberOfActors = this->SlicePlaneActors.size();
 // 	vtkstd::vector<vtkQuadricLODActor*>::iterator
   vtkstd::vector<vtkActor*>::iterator SlicePlaneActorsIterator
 	  = SlicePlaneActors.begin();
